+++ pymc3/ode/__init__.py
@@ -1,0 +1,2 @@
+from . import utils
+from .ode import DifferentialEquation